INSTALL.md: ordered lists can be any number
Markdown properly numbers ordered lists. The main benefit here is that
inserting an element into the ordered list does not require one to
renumber all subsequent elements as is done in issue #15.

To reinforce this habit, renumber every element in the ordered list to
"1", so future edits will resist the urge to unnecessarily renumber
them.

Signed-off-by: Mike Turquette <mturquette@linaro.org>